+++ src/components/ui/card.tsx
@@ -1,11 +1,11 @@
-import React from 'react';
+import React from 'react'
 
 interface Props {
-  children: React.ReactNode;
+	children: React.ReactNode
 }
 
-const Card = ({ children }: Props) => {
-  return <div className="card">{children}</div>;
-};
+const Card = ({children}: Props) => {
+	return <div className='card'>{children}</div>
+}
 
-export default Card;
+export default Card
