+++ src/components/ui/divider.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React from 'react'
 
 const Divider = () => {
-  return <div className="divider" />;
-};
+	return <div className='divider' />
+}
 
-export default Divider;
+export default Divider
