+++ src/components/ui/error.tsx
@@ -1,7 +1,7 @@
-import React from "react"
+import React from 'react'
 
-const ErrorText = ({ children }: any) => {
-  return <div className="error">{children}</div>
+const ErrorText = ({children}: any) => {
+	return <div className='error'>{children}</div>
 }
 
 export default ErrorText
