+++ src/components/ui/toast.tsx
@@ -1,7 +1,7 @@
-import React from "react"
+import React from 'react'
 
-const Toast = ({ children }: any) => {
-  return <div className="toast">{children}</div>
+const Toast = ({children}: any) => {
+	return <div className='toast'>{children}</div>
 }
 
 export default Toast
